feat: Restructure files
Move CLI files to lib/cli
Move builder to builder folder
Change scripts for installation and fix paths in cli.js
