+++ .gitignore
@@ -2,4 +2,4 @@ node_modules
 yarn-error.log
 vendor
 .idea
-cli/deps
+deps
