--- cli/wasm-to-bytes.d.ts
@@ -1,1 +1,0 @@
-export {};
