+++ lib/cli/cli.d.ts
