+++ lib/cli/post-install.js
