--- tsconfig.cli.json
@@ -1,24 +1,0 @@
-{
-  "compilerOptions": {
-    "esModuleInterop": true,
-    "lib": ["es2015", "esnext", "dom"],
-    "module": "esnext",
-    "target": "es2020",
-    "moduleResolution": "node",
-    "alwaysStrict": true,
-    "outDir": "./cli",
-    "declaration": true,
-    "preserveSymlinks": true,
-    "preserveWatchOutput": true,
-    "pretty": false,
-    "forceConsistentCasingInFileNames": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitAny": false,
-    "noImplicitReturns": true,
-    "noUnusedLocals": true,
-    "experimentalDecorators": true,
-    "resolveJsonModule": true,
-    "allowJs": true
-  },
-  "include": ["src/cli/**/*"]
-}
