+++ tsconfig.json
@@ -23,7 +23,9 @@
   "files": [
     "src/index.ts",
     "src/build-tools/near-bindgen-exporter.ts",
-    "src/build-tools/include-bytes.ts"
+    "src/build-tools/include-bytes.ts",
+    "src/cli/cli.ts",
+    "src/cli/post-install.ts"
   ],
   "exclude": ["src/cli/**/*"]
 }
